feat: created and implemented useNoteDetails hook
